refactor: split out macros
